requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.7
gunicorn==23.0.0
packaging==25.0
psycopg==3.2.10
psycopg-binary==3.2.10
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.15.0
